requirements.txt
Flask
pydub
google-generativeai
python-dotenv